Set up the server part that interacts with the KV databases
Create `server.ts` file in the preload package and inside this
file create three functions,

- `openServer` function which is meant to be called by the renderer
part (the frontend) to open up an HTTP server that interacts with the
targeted Deno KV database on behalf of the renderer part.
This function uses `@denokv-gui-client/bridge-server` jsr package
to spin up the HTTP server that provides an API to interact with the
targeted Deno KV database.

- `closeServer` function which is used to close the spun up HTTP
server by `openServer` function.

- `getServerClient` function to export the client that interact
with the spun up HTTP server by `openServer` function.
The client library that this `getServerClient` function exports
is what the renderer package (frontend) is supposed to use
to interact with the targeted Deno KV database.