Allow Kv store's access token to be updated or cleared
Edit the `KvStore` type to make `accessToken` field explicitly nullable
(`string | null`) instead of nullable and optional in the same time. and

Update the SQL query that updates kv stores to allow the access
token column to be set to null by removing the usage of `COALESCE`
function that prevents a null value to be set. And accordingly, Edit the
`KvStore` type to make `accessToken` field required to be set either to
null or string instead of bening optional.