Create `CodeEditor` component with syntax highlighter and formatter
The purpose of this component is to provide a way to write javascript native
values for the keys and values that are used in the Deno KV Database.

The code editor itself of this component built with "codejar" package.
Code formatting is done using "js-beautify" package.
For syntax highlighting, Uses "highlight.js" package + a customized copy
of the "github-dark" theme.