cd-cd: fix landing page deployment workflow
Change the trigger for the landing page deployment workflow to run on
completion of the "Release" workflow instead of "release published" event
which will never trigger the workflow because releases are published from
another GitHub Actions workflow and releases from a workflow can't trigger
release events for other workflows.

And add a condition to the build job to prevent it from execution if
the "Release" workflow didn't complete with successful conclusion.