fix: prevent infinite loading state when network fails or bridge server is unreachable
When a user opens a KV store successfully (network is fine, bridge server is up),
and after that if suddenly network issue occured or the bridge server is no longer
reachable for whatever reason, app will be stuck in infinite loading state if the
user tried to perform any kind of operation (e.g., adding new entry or fetching
new entries). To resolve that, upgrade "@denokv-gui-client/bridge-server"
package to the latest version which contains the fix.

Refactor `fetchEntries` in `kvEntriesState.svelte.ts` to meet the new way
"@denokv-gui-client/bridge-server" package's server returns the requested
entries and the cursor of the next entries.