secure: enable token-based accees control for bridge servers
Previously, when a bridge server is created inside the app to enable
the front-end side (renderer) from accessing Deno kv databases data,
the bridge server is created as a normal http server listening to a
specific port, which is acceessible from anywhere in the host machine,
meaning that any process from outside the app can access the server
if it found the proper port to send requests to.

To close this security risk, first upgrade "@denokv-gui-client/bridge-server"
package that is responsible for creating the bridge servers to version 5,
which comes with token-based accees control feature.

Then use the token-based acceess control feature to protect any
bridge server created inside the app from being accessed from outside
the app.

Also add the option to set an auth token when creating connections to
external bridge servers that are protected with a token-based accees control.