chore: move the development of the bridge server to this repository
Previously, the bridge server (`@denokv-gui-client/bridge-server` package)
was being developed and released separately in
`https://github.com/AbdulrhmanGoni/denokv-bridge-server` repository.

Now I am moving the development of the bridge server to this repository
to be able to maintain it alongside with the other parts of the app in one
place.