build: update manually included dependencies to the final build
- Move "dbmate" from `devDependencies` to `dependencies` to be
able to automatically be included into the final build by `electron-builder`

- Stop copying "hono", "@hono/node-server", "serialize-javascript"
and "dbmate" dependencies manually to the final build in
`electron-builder.mjs` config file because they are now direct
dependencies of a package (bridge-server) in this monorepo and will
be automatically copied into the final build by `electron-builder`.

- Let "@deno" and "@dbmate" modules in the list of files included
manually to be copied into the final build by `electron-builder`
because they contain dynamically imported modules based on the
OS and the architecture, and electron-builder will ignore such
dynamic imports if we didn't tell electron-builder to include them.
Also let `randombytes` module in the list because it's imported
dynamically by "serialize-javascript" package.